requirements.txt
pandas
streamlit
pillow
tensorflow
numpy